+++ src/Traits/PdfRotationTrait.php
@@ -72,8 +72,8 @@ trait PdfRotationTrait
         $angle *= \M_PI / 180.0;
         $cos = \cos($angle);
         $sin = \sin($angle);
-        $cx = $x * $this->scaleFactor;
-        $cy = ($this->height - $y) * $this->scaleFactor;
+        $cx = $this->scale($x);
+        $cy = $this->scale($this->getPageHeight() - $y);
         $this->outf(
             'q %.5F %.5F %.5F %.5F %.2F %.2F cm 1 0 0 1 %.2F %.2F cm',
             $cos,